Cast epoch time to integer

+++ batch_jobs/cassandra_stock_twits/add_stock_twits.py
@@ -21,7 +21,7 @@ import datetime
 epoch = datetime.datetime.utcfromtimestamp(0)
 
 def unix_time_millis(dt):
-    return (dt - epoch).total_seconds() * 1000.0
+    return int((dt - epoch).total_seconds() * 1000)
 
 
 twits_location = "twits"
